New UI test related project structure
