--- oscm-portal-integrationtests/.gitignore
@@ -1,3 +1,0 @@
-/target/
-
-
