+++ oscm-ui-tests/src/main/java/org/oscm/webtest/AppHtmlElements.java
@@ -1,0 +1,28 @@
+/*******************************************************************************
+ *
+ *  Copyright FUJITSU LIMITED 2018                                           
+ *
+ *  Creation Date: 20 6, 2018                                                      
+ *
+ *******************************************************************************/
+package org.oscm.webtest;
+
+public class AppHtmlElements {
+    public static final String APP_CONFIG_DIV_CLASS_STATUS_MSG = "statusPanel";
+    public static final String APP_CONFIG_LICLASS_STATUS_MSG_OK = "statusGreen";
+    public static final String APP_CONFIG_LICLASS_STATUS_MSG_ERROR = "statusRed";
+
+    public static final String APP_CONFIG_FORM1 = "configurationSettings";
+    public static final String APP_CONFIG_FORM1_INPUT_END_NEWCONTROLLERID = ":newControllerId";
+    public static final String APP_CONFIG_FORM1_INPUT_END_NEWORGID = ":newOrgnizationId";
+    public static final String APP_CONFIG_FORM1_INPUT_END_EXISTORGID = ":configurationValue";
+    public static final String APP_CONFIG_FORM_BUTTON_CLASS = "oscm_app_button";
+
+    public static final String APP_CONFIG_FORM2 = "appSettings";
+
+    public static final String APP_SERVICEINSTANCE_TITLE = "APP Sample Controller Configuration";
+    public static final String APP_SERVICEINSTANCE_TABLE_ID = "serviceForm:serviceTable";
+
+    public static final String APP_SAMPLECONTROLLER_FORM_ID = "configurationSettings";
+
+}
