+++ oscm-ui-tests/src/main/java/org/oscm/webtest/exception/ConfigurationException.java
@@ -1,0 +1,15 @@
+/*******************************************************************************
+ *
+ *  Copyright FUJITSU LIMITED 2020
+ *
+ *  Creation Date: 10.01.2020
+ *
+ *******************************************************************************/
+package org.oscm.webtest.exception;
+
+public class ConfigurationException extends RuntimeException {
+
+    public ConfigurationException(String message) {
+        super(message);
+    }
+}
