+++ oscm-ui-tests/src/main/resources/TechnicalService.xml
@@ -1,0 +1,41 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<!-- Copyright FUJITSU LIMITED 2018 -->
+
+<tns:TechnicalServices xmlns:tns="oscm.serviceprovisioning/1.9/TechnicalService.xsd"
+                       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                       xsi:schemaLocation="oscm.serviceprovisioning/1.9/TechnicalService.xsd ../../../../oscm-serviceprovisioning/javares/TechnicalServices.xsd">
+
+    <tns:TechnicalService id="AppSampleService" build="2018.07.20" provisioningType="ASYNCHRONOUS"
+                          provisioningUrl="http://oscm-app:8880/oscm-app/webservices/oscm-app/oscm-app/org.oscm.app.v2_0.service.AsynchronousProvisioningProxy?wsdl"
+                          provisioningVersion="1.0"
+                          accessType="DIRECT" loginPath="Dynamically provided." baseUrl="">
+
+        <AccessInfo locale="en">Description of how to access the application.</AccessInfo>
+
+        <LocalizedDescription locale="en">Sample APP implementation.</LocalizedDescription>
+        <LocalizedDescription locale="de">Sample APP Implementierung.</LocalizedDescription>
+        <LocalizedLicense locale="en"/>
+
+        <ParameterDefinition id="APP_CONTROLLER_ID" configurable="false" mandatory="true" valueType="STRING"
+                             default="ess.sample">
+            <LocalizedDescription locale="en">The ID of the sample controller implementation.</LocalizedDescription>
+        </ParameterDefinition>
+        <ParameterDefinition id="PARAM_MESSAGETEXT" configurable="true" mandatory="true" valueType="STRING"
+                             default="Some message.">
+            <LocalizedDescription locale="en">The message text for emails.</LocalizedDescription>
+        </ParameterDefinition>
+        <ParameterDefinition id="PARAM_EMAIL" configurable="true" mandatory="true" valueType="STRING"
+                             default="user@myorg.com">
+            <LocalizedDescription locale="en">The receiver of emails.</LocalizedDescription>
+        </ParameterDefinition>
+        <ParameterDefinition id="PARAM_USER" configurable="true" mandatory="true" valueType="STRING"
+                             default="defaultuser">
+            <LocalizedDescription locale="en">IAAS user</LocalizedDescription>
+        </ParameterDefinition>
+        <ParameterDefinition id="PARAM_PWD" configurable="true" mandatory="true" valueType="PWD"
+                             default="">
+            <LocalizedDescription locale="en">IAAS password</LocalizedDescription>
+        </ParameterDefinition>
+    </tns:TechnicalService>
+
+</tns:TechnicalServices>
