+++ oscm-ui-tests/src/main/resources/webtest.properties
@@ -1,0 +1,25 @@
+bes.https.url=https://localhost:8081
+bes.http.url=https://localhost:8081
+bes.secure=false
+bes.user.id=administrator
+bes.user.password=admin123
+bes.user.key=1000
+app.https.url=https://localhost:8881
+app.http.url=https://localhost:8881
+app.user.id=administrator
+app.user.password=admin123
+app.secure=false
+email.address=bes.estfarmaki@dev.est.fujitsu.com
+email.host=estdevmail1
+email.user=bes.estfarmaki@dev.est.fujitsu.com
+email.password=secret
+email.protocol=imap
+email.createaccount.head=Account created
+email.createaccount.phrase.password=Your initial password is:
+email.createaccount.phrase.userkey=User key for Web service access:
+technicalservice.xml.path=/build/oscm-ui-tests/src/main/resources/TechnicalService.xml
+technicalservice.name=AppSampleService
+create.subscription.count=10
+create.subscription.waiting.seconds=10
+auth.mode=OSCM authentication mode (INTERNAL or OIDC)
+chrome.driver.path=/usr/local/bin/chromedriver
