+++ oscm-ui-tests/src/test/java/org/oscm/portal/PortalUserTest.java
