--- oscm-webtest-base/.gitignore
@@ -1,3 +1,0 @@
-/target/
-
-
