--- oscm-webtest-base/readme.md
@@ -1,4 +1,0 @@
-*Webtest Base*
-
-Run mvn clean package install
-
