--- oscm-webtest-base/src/main/java/org/oscm/webtest/AppPathSegments.java
@@ -1,15 +1,0 @@
-/*******************************************************************************
- *                                                                              
- *  Copyright FUJITSU LIMITED 2018                                           
- *                                                                                                                                 
- *  Creation Date: 20 6, 2018                                                      
- *                                                                              
- *******************************************************************************/
-package org.oscm.webtest;
-
-public class AppPathSegments {
-    public static final String APP_CONFIGURATION = "/oscm-app/default.jsf";
-    public static final String APP_SAMPLE_CONTROLLER = "/oscm-app-sample/";
-    public static final String APP_SERVICE_INSTANCE = "oscm-app/controller/?cid=";
-
-}
