--- oscm-webtest-base/src/main/java/org/oscm/webtest/PortalPathSegments.java
@@ -1,31 +1,0 @@
-/*******************************************************************************
- *                                                                              
- *  Copyright FUJITSU LIMITED 2018                                           
- *                                                                                                                                 
- *  Creation Date: 20 6, 2018                                                      
- *                                                                              
- *******************************************************************************/
-package org.oscm.webtest;
-
-public class PortalPathSegments {
-
-    public static final String CREATE_MARKETPLACE = "shop/createMarketplace.jsf";
-    public static final String DELETE_MARKETPLACE = "shop/deleteMarketplace.jsf";
-
-    public static final String CREATE_ORGANIZATION = "operator/createOrganization.jsf";
-
-    public static final String IMPORT_TECHNICALSERVICE = "techservice/import.jsf";
-    public static final String UPDATE_TECHNICALSERVICE = "techservice/edit.jsf";
-
-    public static final String DEFINE_MARKETSERVICE = "service/add.jsf";
-    public static final String DEFINE_PREICEMODEL = "priceModel/servicePriceModel.jsf";
-    public static final String DEFINE_PUBLISHOPTION = "service/publish.jsf";
-    public static final String ACTIVE_MARKETSERVICE = "service/activation.jsf";
-
-    public static final String REGISTER_CUSTOMER = "organization/addCustomer.jsf";
-
-    public static final String GOTO_MARKETPLACE = "shop/gotoMarketplace.jsf";
-    public static final String INDEX_MARKETPLACE = "index.jsf";
-    public static final String URL_MARKETPLACE_ID = "index.jsf?mId=";
-
-}
