--- oscm-webtest-base/src/main/resources/log4j.properties
@@ -1,26 +1,0 @@
-# Copyright FUJITSU LIMITED 2018
-# Root logger option
-log4j.rootLogger=OFF
- org.apache.commons.logging.Log=org.apache.commons.logging.impl.NoOpLog
-log4j.logger.org.apache=OFF
-log4j.logger.httpclient=OFF
-log4j.logger.org.oscm.webtest=DEBUG, FILE
-log4j.logger.org.oscm.portal=DEBUG, FILE
-# Direct log messages to stdout
-#log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-#log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-#log4j.appender.stdout.layout.ConversionPattern=%d{ABSOLUTE} %5p %c{1}:%L - %m%n
-
-# CONSOLE is set to be a ConsoleAppender using a PatternLayout.
-log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
-log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
-log4j.appender.CONSOLE.layout.ConversionPattern=%d [%t] - %-5p - %m%n
-
-
-# Rolling File Appender system
-log4j.appender.FILE=org.apache.log4j.RollingFileAppender
-log4j.appender.FILE.File=../shippable/logs/webtest.log
-log4j.appender.FILE.MaxFileSize=10MB
-log4j.appender.FILE.MaxBackupIndex=5
-log4j.appender.FILE.layout=org.apache.log4j.PatternLayout
-log4j.appender.FILE.layout.ConversionPattern=[%d{MM/dd HH:mm:ss}] [%t] %-5p %m%n
